requirements.txt
flask
flask-cors
google-generativeai
pandas
openpyxl
python-dotenv
pillow
